put environment classes into a new file
