+++ src/interpreter/main.py
@@ -1,69 +1,19 @@
 """AST interpreter"""
 
 
-from dataclasses import dataclass
-from typing import Optional
-
 from lark.tree import Meta
 
 from bl_ast import nodes
 from bl_ast.base import ASTVisitor
 
 from . import values
-from .base import Result, ExpressionResult, BLError, \
-                  error_not_implemented, error_var_nonexistent
-from .values import Value
+from .base import Result, ExpressionResult, Success, BLError, \
+                  error_not_implemented
+from .values import Value, PythonFunction
+from .env import Env
 
 
 #pylint: disable=too-many-return-statements
-
-
-@dataclass
-class Var:
-    """Interpreter mutable binding"""
-
-    value: Value
-
-
-class Env:
-    """Interpreter environment"""
-
-    vars: dict[str, Var]
-    parent: Optional['Env']
-
-    def __init__(self, parent: Optional['Env'] = None):
-        self.vars = {}
-        self.parent = parent
-
-    def new_var(self, name: str, value: Value) -> None:
-        """Set a new/existing variable"""
-        self.vars[name] = Var(value)
-
-    def get_var(self, name: str, meta: Meta) -> ExpressionResult:
-        """Retrieve the value of a variable"""
-        resolve_result = self.resolve_var(name, meta)
-        match resolve_result:
-            case Var(value=value):
-                return value
-            case BLError():
-                return resolve_result
-
-    def set_var(self, name: str, value: Value, meta: Meta) -> Optional[BLError]:
-        """Assign to an existing variable name"""
-        resolve_result = self.resolve_var(name, meta)
-        match resolve_result:
-            case Var():
-                resolve_result.value = value
-            case BLError():
-                return resolve_result
-
-    def resolve_var(self, name: str, meta: Meta) -> Var | BLError:
-        """Resolve a variable name"""
-        if name in self.vars:
-            return self.vars[name]
-        if self.parent is not None:
-            return self.parent.resolve_var(name, meta)
-        return error_var_nonexistent.fill_args(name).set_meta(meta)
 
 
 class ASTInterpreter(ASTVisitor):
@@ -124,7 +74,7 @@ class ASTInterpreter(ASTVisitor):
                 elems = []
                 for e in elems_in_ast:
                     e_visited = self.visit_expr(e)
-                    if not isinstance(e_visited, values.Value):
+                    if not isinstance(e_visited, Value):
                         return e_visited
                     elems.append(e_visited)
                 return values.BLList(elems)
